fix spurious missing-nul-check on free for memaligned memory
normally zero check bytes are never written except at group init, but
when producing aligned results via a nominal chunk start at an offset
for alignment, we can't assume there's already a check byte in place.
so add it here.

the underlying issue is really that the check should not be performed
at -4 relative to the nominal allocated object address but rather -4
relative to the entire slot. as-is, we're only catching overflows that
extend all the way to the used portion of the next slot. fixing this
will take more work, but it's not critical to correct operation, only
an improvement to catch more errors in the caller.