fix use of stale mask in nontrivial_free
mask could have changed in a way that impacts the code path that needs
to be taken in nontrivial_free, between the failure of the cas loop in
free and obtaining the write lock. don't try to do anything fancy
here; just load it properly after taking the lock. this simplifies
things too.