fix format string error in print_group debug function
%d was correct at some point before support for printing sizes of mmap
pseudo-groups was added. now the correct type for the expression is
uintptr_t, but some compilers oddly (wrongly) treat the bitfield as
its own narrower type. improve readability and avoid the need for a
specifier for uintptr_t and erroneous warnings by just using an
intermediate variable of type size_t, the correct semantic type here.