harden groups allocated inside another group's slot against invalid free
one of the disadvantages/dangers of allocating small-count, small
size-class groups inside a slot in a larger-size group is that the
application could still have a pointer from a previous use of the
outer memory slot and, through a double-free bug, free the memory
inside which other allocations live. there doesn't appear to be a
low-cost way to preclude this in the safe out-of-band metadata, but we
can sabotage attempts to free prematurely with invertible corruption
to the outer slot's in-band metadata.

there are lots of ways we could do this, but for now, just add 6 to
the reserved-bytes count at offset -3. the only values valid in this
field are 0-5; larger counts are encoded in a 32-bit int at the end of
the slot. adding 6 ensures that the value is invalid, and is easy to
back out at free_group time.

later, when we have secure random numbers, it may be preferable to
corrupt the offset field in a way the application can't easily
falsify to invert to a valid offset.