move mmap call for large allocations outside of lock
ultimately, all mmap and munmap calls except for meta areas should be
moved to take place without lock held, to avoid delaying allocations
that can be serviced immediately with existing groups. doing this for
multi-slot groups is complex and can lead to over-allocation if not
done right, since other threads need to be able to see that an
allocation is already being made and avoid performing their own.
however the case of directly mmap-serviced allocations is easy; the
only reason this code path needs a lock at all is to obtain a metadata
slot to use.