fix dump_heap debug output not to show full but active groups twice
it's possible for a group to remain active despite having no free or
available slots, due to malloc taking the fast path and just
atomically removing the last availability bit. such groups were
printed twice, both on the inactive groups list and their size class's
active list.

not being on a list should be the condition for printing as inactive.