remove 4k minimum alloc_group heuristic
inspired by buddy-allocator-like results from lowering the limit from
PAGESIZE to 4k and testing with simulated large pages. for example, an
initial malloc(10) would previously start by allocating an 8x496 group
to hold the 30x16 group. now, it recursively splits a 4k page to
contain 2x2032, containing 2x1008, containing 2x496, containing the
desired 30x16. this leaves more flexibility for how the remaining
slots of size 2032, 1008, and 496 can be used, reducing the likelihood
of allocating another page for a program with minimal dynamic memory
utilization.

of course this recursive splitting takes a little more time, and
produces smaller groups which will need to be cycled sooner, and in
general, flexibility of use can be bad, leading to more complex
fragmentation. however, only a small finite number of these low-count
groups can exist at any given time. once usage in the size class
passes a certain threshold, all further group allocations will be
full-count, and for the size classes handled by this code path (class
8 and up, i.e. size 144 and up), full count is at least 4k.

as an aside, the idea of this heuristic (which is now being removed)
goes back to before it was determined that choices of counts for each
size class should be quantized in a way that ensures all groups have
size approximately (but not exceeding) a power of two. when there was
no clear structure on how smaller-size-class groups would end up being
allocated inside larger ones, the idea of avoiding that seemed very
appealing.