use coarse size classes for early allocations in class
the idea here is that even size classes have minimum counts of 5 or 7,
compared to 2 or 3 for odd size classes, and we'd like to start off
with very small counts so as not to impose unnecessary growth on
programs that only perform a very small amount of allocation of a
particular size class.

the logic has to be here in malloc, not in alloc_group, because we
want to use the coarse class in cases where it already has an active
group with available slots.

this seems like it's wasting memory, and like it would be simpler to
just waste the memory by allowing allocation of even classes with
counts of 2 or 3, leaving half a slot unusable. however, by instead
moving up to the next class, we allow the groups to be shared with
allocations of the next size up.

incidentally, the FIXME comment removed here was intended to refer to
adding some sort of coarse-size-classing logic.