fix incomplete double-free/use-after-free checks
a free slot may be reflected in avail_mask rather than freed_mask if
it's cycled around to the active group, so the existing check in
free() was incomplete. moreover, realloc and malloc_usable_size
completely failed to check that the memory passed to them was not
already freed.

putting an added check for avail_mask in the cas loop of free was
considered, but it adds cost in a critical code path (possibly
increasing the window for cas to fail), and is not significantly
stonger than checking avail_mask once in advance. by putting the check
in get_meta, all functions which lookup the meta automatically
benefit, without the need for function-specific logic.

the old check is left in the cas loop since it's inexpensive, being
that we have to load freed_mask in the loop anyway, and hitting just
an invalid concurrent free is more plausible than hitting both a
concurrent invalid free and a cycling of the containing group to the
active position for its size class.