remove redundant variable m in malloc mmap code path
this was an artifact of the old group variable name being "cur", which
didn't make sense outside the context of being the active group for a
size class.

style is generally that g is used to represent a reference to a group
via its metadata, and m is used to represent a reference to the
metadata itself. here, we're creating a classless length-1 group, so g
is appropriate.