begin separation into multiple files
all types and helper functions for working with metadata are moved
into meta.h, functions marked with inline to suppress unused warnings.
the dump_heap debug functionality is not yet split because it depends
on global state and lock primitives that aren't yet setup to be usable
outside of malloc.c.