move lock definitions out of malloc.c to a new header
this facilitates splitting functions out into separate files, and sets
up to have a more convenient place to swap different lock designs. in
order to do the latter, though, some changes to the abstraction will
be needed.

ideally the lock would live inside the malloc_context, but making that
works creates a circular dependency between locking.h and meta.h,
which needs to be resolved somehow.