fix inability to free last group of a class without getting lucky
the logic in free's fast path to bail out and take the lock, which is
needed to free the group, looked for a full bitset in freed_mask. this
precluded freeing of an active group with nonzero avail_mask.

the scenario this change is designed to handle is the case where,
after heavy usage in a size class, all groups or at least the last
active group are large-count (possibly up to 1MB in total size), and
the program has no further use for objects of that particular size
class. in this case, the group would stick around consuming memory.

to implement this change, a relaxed access to avail_mask is made in
free's cas loop, and used in computing whether the combination of
avail_mask and freed_mask is a full bitset. since the access is
relaxed, we may see a stale value. failure to see removal of bits from
avail_mask (new allocations) is inconsequential; at worst it produces
a false positive and we see the correct value in nontrivial_free after
taking the lock. the other possible case where we see a stale value is
when try_avail has cycled the group to active. in this case it seems
possible to miss a chance to free the group, seeing both the new
(zero) value of freed_mask and the old (zero) value of avail_mask. it
may be possible to mitigate this with a barrier (relying on the
existing one between updates on the try_avail side), but that's
probably overly costly for the sake of preventing what should be
nothing but an extremely rare missed chances to free a group.

this commit exacerbates an existing problem with unwanted freeing of
small groups that could resonably kept around in anticipation of
future use, which will be fixed separately.