add logic for deciding whether to free group when last slot is freed
previous changes to allow freeing of an active group with available
slots exposed a long-known problem, whereby certain usage patterns
could produce heavy cycling between mmap and munmap when it would be
preferable to keep the memory ready to use for future allocations.
however, this always involves tradeoffs between memory usage,
fragmentation, and performance.

this commit introduces logic to mitigate excess cycling of mappings,
with bounds on the potentially-wasted memory. it favors freeing of
groups larger than what alloc_group would allocate for the class based
on current usage (thereby bounding relative space cost) and of "old"
groups with low slot count in classes with high usage. it always frees
when the group is not the last one with available slots in its class,
and always frees groups allocated inside a slot of a larger group
(which recursively tie up and fragment memory), but otherwise avoids
freeing small-size-class groups (for which absolute space cost is
low).