record and cycle used offset within a slot each time it's allocated
the allocator's data structures have never required the address
returned to the application to lie at the beginning of the allocated
slot. previously, the allowance that it be offset within the slot was
utilized only for memalign, and the offset used was not recorded
anywhere discoverable, only implicitly in the in-band header which is
validated against OOB metadata. this precluded iteration of the entire
heap for debugging or instrumentation.

however, from the beginning the mechanism for allowing allocations to
start at an offset within their slot was intended to admit cycling or
randomizing the offset, to increase the capability to catch
double-free errors. double-free is already detectable and trappped
when the freed slot is still unused at the time of the second call to
free, but a fundamental property of of reusable resource identifiers
is that it's impossible to distinguish a double-free of an old
reference to the freed resource from an intended free of the new
resource whenever the exact same resource identifier has been handed
out again to refer to a new resource. however, by changing the offset,
it's possible to reuse a slot without assigning an identical resource
identifier right away.

the cycle length possible varies greatly by size class and by how
closely the requested size matches a size class (more internal
fragmentation gives longer cycles until reuse). in principle, a
hardening-focused build option could always request a size class
appropriate for n+16*(k-1) bytes to ensure a minimal cycle length of
k, at the expense of significant memory expenditure at small sizes.