validate meta structes via random secret in meta_area header
something like this was always intended. previously, the only
validation of the meta object obtained by following the slot offset
back to the group header and the group header back to the OOB meta was
that the candidate meta object had a matching pointer back to the
group header. if an attacker could construct such a pair of matching
pointers and cause a slot from the falsified (or corrupted in-place)
group to be freed, they could inject the meta object into the free
list, and cause malloc to hand out memory they control, or cause it to
engage in (perhaps limited) target overwrite of pointed-to data.

by requiring meta objects have a check value in the corresponding meta
area header that matches a global secret, such an attack becomes
impossible without having a means to read either the the secret or the
check value from an existing meta_area header.

getentropy backend to obtain random bits for the secret is hard-coded
for now. for upstream integration it will hook into a userspace csprng
with full fallback coverage.