cycle offsets in individually-mmapped allocations
we have no persistent state for the addresses associated with these,
which the kernel will happily reuse on subsequent mmaps, so instead
keep a global counter and reduce it modulo size of internal
fragmentation (from rounding up to a multiple of 4k). no care for
exactness needs to be taken here since the candidate value is just
injected as the "previously used offset", which enframe() will use to
safely derive a next offset to use.

on archs with larger page size, longer intervals to reuse could be
achieved (and larger in-place realloc without mremap could be
supported) by computing maplen accurate to full page granularity
rather than units of 4k; such an enhancement could be considered for
later.