cache page size in malloc context if it's runtime-variable
up to now, all the uses of page size have been in non-bottleneck code
paths so potentially calling sysconf or equivalent to lookup the page
size has not been a problem. however not having it available for
efficient access has been an obstacle to development of functionality
where it may be needed in fast paths.