release physical memory for page-aligned subranges on free
this applies only to allocations large enough to contain at least a
full (aligned) page, and skips others with no significant work. slots
in a group are not themselves page-aligned, so the first and last page
overlapping the the slot generally can't be released. in principle
they could be when the adjacent slot is also free, but that would
require further synchronization and madvise under lock, and would also
preclude having persistent state to cycle used offset within the slot.

for now, MADV_DONTNEED is used to release pages of physical memory.
MADV_FREE would be faster and may be the right thing to switch to
later, but would make it harder to observe the effectiveness/benefits
of the change since it's not reflected in memory usage until memory
pressure causes the pages to be discarded. MADV_DONTNEED also
potentially allows making "whole-page subranges always start
zero-filled" an invariant that calloc could use (and that calloc and
possibly malloc could assert to catch some use-after-free).