fix race condition in lock-free path of free
the observed value of avail in the lock-free cas loop is only reliably
an upper bound on the actual value of the group's avail_mask. if a
concurrent call to malloc removes the last available slot and
therefore removing the group from the active list, free could fail to
see that it's transitioning the group from full to non-full state and
fail to add it back to the active list. this prevents it from being
used to satisfy future allocations, and if the group ever becomes
entirely freed, it will crash in nontrivial_free due to violation of
the invariant that groups with free slots are always on a list.

to avoid the race, use only the freed_mask where the combined
freed|avail mask was being used to determine if the group may be
transitioning from full to non-full state. this will take the slow
(locking) path in some cases where it's not necessary (the first free
in a group after it becomes the active group for its size class), but
there do not seem to be other solutions without more expensive
tradeoffs.