fix spurious madvise of size 0 on some frees
the xor expression for checking that a slot crosses two page
boundaries can have false positives if it crosses a larger
power-of-two address space boundary. in this case the rounding up to
page boundary operation is still valid, but end-base is less than a
full page and results in a no-op syscall with size zero.

for now, rather than impacting the hot path, just do a second check
for nonzero length before making the syscall.