add bounce counter to cut off perpetual cycling of same mmap/munmap
the ok_to_free function in free.c decides whether to free an empty
group or keep it around for future use based on a number of criteria.
in large programs with lots of malloc usage, these criteria should
provide decent guard against getting in a state of rapidly tipping
back and forth between allocating and freeing the same group. however,
with low usage, for size classes 24 and up (sizes over 2028 bytes),
it's easy to get in a situation where the same group is mapped and
unmapped over and over again, yielding pathological performance.

mitigate this with a one-byte bounce counter that cuts off ok_to_free
after 255 cycles. this avoids penalizing small programs that barely
use malloc with permanent mappings while bounding the amount of time
wasted. it's not tuned and probably could be improved significantly,
but it's simple and works.