track usage by class as a slot count rather than aggregate size
this simplifies/optimizes the code paths by removing table lookups and
muls. the aggregate size was not really accurate anyway since it
didn't account for overhead, and would become even more misleading
once support for individual mmaps below the hard mmap threshold is
added.

it's intended that this commit not produce any functional changes. it
required rearranging some algebraic expressions where division by 2
was exact when used on size (due to a known factor of 16) but not on
counts, which may be odd.