increase the threshold under which coarse size classes are used
the idea here is to bound the size of the new group to be created
relative to the existing coarse-class usage. if usage is 7 (just above
the old threshold), the new fine-class group will have 5 or 7 slots,
nearly doubling total usage. if usage is 13 (just above the new
threshold), the new group will be at most half the size of the current
usage.