prepare free for single-slot groups with a size class
for groups with more than one slot, it's impossible to have both a
transition from full to non-full and partly-in-use to all-free in the
same call to free. this is just the fact that the mask has more than
one bit. single-slot groups used for classless, direct-mmapped
allocations had a special case by their size class to skip
deactivation logic, but in preparation for single-slot groups in size
classes, we need to actually look at whether the group is on an active
list.

it's an invariant-violation for classless groups to be on an active
list, so keep the check as an assertion.