don't force munmap of single-slot groups
this is in preparation to support classful single-slot groups. a
single-slot group that does not have a size class must be unmapped on
free (since there is nowhere to track it for reuse), and one allocated
with less space than the nominal class size can't be reused (because
it may be too small).

otherwise, there's nothing invalid about keeping single-slot groups,
so let the remainder of the logic in okay_to_free decide whether to
unmap or not rather than forcing it.

this will allow us to start off with slot count of 1 for certain size
classes -- particularly all of the ones that are just below a power of
two (indices 3 mod 4), as well as some others -- to limit eagar
allocation in programs with low malloc usage.