reduce the minimum slot counts for group allocations serviced by mmap
this change drops the invariant that all group mmaps consist of a
power-of-two number of pages to reduce eagar allocation of slots that
might never be used.

except for class indices 3 mod 4 (sizes just under powers of two)
where 1 is naturally a valid count, the new minimum counts are floors
of 3/8, 3/4, 5/8, or 7/16 of the original minimum counts, and produce
groups that consume 3n, 5n, or 7n pages where n is a power of two.

for each size class, these smaller minimum counts will only be used a
few times before reaching usage levels that warrant larger counts, and
thus the number of irregularly-sized groups is bounded.