ensure compiler can see division by 16 is unsigned
the /16U was intended to ensure that the compiler doesn't miss the
invariant that the numerator is always positive and generate a
costlier signed division (requires fixup after bitshift to truncate
towards zero). however, on 64-bit archs the numerator has higher rank
(a signed 64-bit type) and the 16U just promotes to a signed type.