replace magic number 16 and sizeof(struct group) with UNIT
this change is mainly to make the code clearer and more
self-documenting, but in theory would allow definition of a smaller
fundamental unit/alignment or a larger one if needed. the actual
constraints are that:

- UNIT is a power of two
- sizeof(struct meta *)+4 <= UNIT
- alignof(max_align_t) <= UNIT.

in fact UNIT could be defined as the max of those two quantities
rounded up to a power of two, but doing so would require evaluation of
whether the size classes are sufficiently granular at small sizes with
a smaller unit.