fix unexpected allocation of 7x144 group in non-power-of-two slot
groups allocated nested inside the slot of another group are supposed
to only use (just-below) power-of-two sizes (class indices 3 mod 4).
this avoids having a wide spread of sizes used, potentially requiring
allocation of larger groups that won't be used.

it was intended that all sizes classes 8 and up fit a divisibilty
pattern that allows use of med_cnt_tab to select the number of slots
to use. however, in the case of class 8 (size 144), following the
pattern and using 7 for the count produces exactly 1008 bytes (63
units) and does not leave room for the 16-byte (1 unit) group header.
thus, these groups were being allocated in slots of the next size up,
1168, with a minimum group size of 7 slots (8k). this resulted in
significant overhead for allocating a single object of size 125-140.

this can be solved by extending small_cnt_tab by one element and only
using the med_cnt_tab pattern starting at class 9. for now, the
minimal count is just dropped from 7 to 6, but this leaves significant
unused space. it would probably be better to introduce coarse size
classing for small size classes, and start with 14 slots as soon as we
switch to fine classing, but that requires a change with more
extensive impact that should be evaluated separately.