limit slot count growth to 25% instead of 50% of current usage in class
this may turn out not to be the right decision or require further
tuning, but it makes a significant reduction in overall memory usage
by small programs. reduction was 15% in one test case based on
real-world application traces, and could be as much as 25% in
worst-case allocation-only, or more with an unlucky mix of frees.

because slot counts also grow exponentially up to the maximum for the
class, this change does not cause an explosion in number of low-count
groups; it only doubles the number at each slot count, thereby
doubling the total possible number in existence at any given time.