adjust several size classes to fix nested groups of non-power-of-2 size
analogous to a709dde9aca71e1959e389f1cfea974285f4f5ce. all of these
were arithmetic errors in the original design of the size class
spacing, but escaped notice without detailed analysis. while I at
first suspected these were the result of only testing divisibility of
the class (sc|3)+8 by 7, 6, or 5 (excluding the 3-slot case), that
only matches the error for classes 13, 21, and 29 (the odd ones). it
looks likely that I also only checked divisibility of whole 4k pages
(non-nested group) for sizes sufficiently large to warrant direct
mmap, assuming 4k pages; that explains the error in classes 18 and 20.
this is actually okay on archs with 4k page size, but it would cause
severe over-allocation on archs with larger pages.

as with the previous issue at class 8, all of these interfere with
ability to do coarse size classing at the size class just above the
slot size they should be using, and cause significant over-allocation.
fix this by reducing the size of the affected classes by one unit.
this produces the following changes, scaled by unit:

class 13: 336 -> 320
class 18: 816 -> 800
class 20: 1168 -> 1152
class 21: 1360 -> 1344
class 29: 5456 -> 5440

this may of course throw some allocations up to a larger size class,
but it's preferable to the alternative. it's possible that some of
this could be recovered by using the old values for the even size
classes contingent on pagesize being constant and sufficiently low.