move coarse size classing logic to malloc fast path
this was originally not done out of concern that it would impact
performance of the common case once usage is at levels where coarse
size classing shouldn't be used. however, the entire logic can be made
conditional on a condition that already has to be checked in the fast
path, whether ctx.active[sc] is null.

prior to this change, use of coarse size classes was somewhat costly
to performance, in that it prevented the malloc fast path from being
usable until usage hit the threshold to switch to fine classing. now,
it should be nearly free, and this expands the possibilities for
using it at smaller sizes.

this commit is intended not to produce any functional change.