allow slot count 1 for size classes 3 mod 4 as natural/first-class
despite naturally admitting count 1 according to divisibility,
previously, count could only be reduced from 2 to 1 in the
reduced-count code path for the mmap case. this prevented use of
single-slot groups after the bounce counter for the size took effect,
thereby preventing the cached mapping from being used to satisfy
future requests for the same size.

instead, treat count 1 as a first-class group length. this simplifies
the loop to choose a count. then, if the chosen count of 1 ends up
being too small to use mmap, increase it to 2, making it eligible for
nesting if it's still not large enough for mmap.

a couple "off by one" issues are fixed at the same time. previously
they were not relevant, since size*cnt+UNIT could never be an exact
power of two when cnt was at least 2. however, with cnt=1 being a
possibility, each size class 3 mod 4 has size+UNIT exactly equal to a
power of two.