adjust definition of stride not to assume single-slot group is mmapped
commit d92266d28a87602977151a4eea35ca6a3d13a47f redefined stride in
terms of maplen for single-slot groups, to allow single-slot mmapped
groups that don't need the full page count to provide their nominal
size. however, we should not assume that all single-slot groups are
backed by mmap. when donation of slack space at beginning/end of data
segments is added as part of integration with musl, there will be
non-freeable, non-mmap-backed single-slot groups. go ahead and adjust
the definition of stride now to prepare for them.