fix off-by-one in logic to use single-slot groups
the boundary case req==size+UNIT is possible only for size classes
just below a power of two (classes 3 mod 4). in this case, the
existing logic caused repeated usage of single-slot allocation until
usage was sufficiently high (32 total slots) to make the next count
exceed 7. while there is some appeal to having that many allocations
be individually freeable, it's probably a bad tradeoff since it causes
fragmentation of the virtual address space and makes it more likely to
activate bounce counters.

with a strict inequality condition, the single-slot group logic will
not trigger for these size classes when no memory would be saved. slot
count of 1 will still be used naturally until usage bumps it to 2.

in cases (only present at higher sizes) where single-slot really does
save memory, a large number (up to 32) of such allocations will still
be used. these are closer to the hard mmap threshold anyway and
probably not concerning, and the savings seem sufficient to warrant
it.