second partial overhaul of bounce counter system
commit 5bff93c8354f76bc031d87b5173bedda70fe3036 was wrong in that
accounting bounces to map sizes rather than size classes just didn't
work -- at free_group time for groups allocated with reduced count or
single-slot (individually mmapped), it was impossible to determine the
map size they "would have been allocated in".

restore tracking of bounces by size class, but keep the new system of
caching mappings by power-of-two class. drop support for caching
multiple mappings of the same size for now, since accounting to do it
well looks to be complicated and it's not clear that it's needed.

there are still significant improvement I'd like to make in this area.