initial conversion of bounce counting to use sequence numbers, decay
simply counting group mmap events with a corresponding previous munmap
necessarily accounts for and accumulates events which may have been
distant from one another in time and unrelated, leading long-lived
programs with continued use of the allocator to eventually start
saving a fully-free group in all classes they use.

this commit is the beginning of a solution, based on keeping a
sequence number incremented on "heavy" events (mmap/munmap) and only
counting bounces that occur with few intervening heavy events. in some
sense this is serving as a time-scale-invariant substitute for simply
measuring closeness in wall-time, but aside from having problems with
being tied to particular execution speed, using wall time would fail
to mitigate unwarranted high relative cpu usage by processes that only
execute in sporadic timeslices but bounce each time they do.

the other major ingredient in this solution is decaying bounce counts
to periodically probe whether retention of fully-free groups is still
needed. as long as the decay rate is sufficiently low, the relative
time cost of the additional munmap/mmap cycles will be low enough not
to matter.

sequence number advancement, bounce threshold, and decay rate all need
further tuning. i'm committing this initial version in a somewhat
messy state so as not to mix refactoring with functional changes, and
to give a basis for further tuning to take place on top of.