fix complete regression (malloc always fails) on variable-pagesize archs
commit 5bff93c8354f76bc031d87b5173bedda70fe3036 moved the call to
alloc_meta later to avoid doing it in the case where a meta struct was
already available from a mapping being reused. that functionality was
reverted but the call to alloc_meta was not moved back, and the PGSZ
macro depends on alloc_meta having been called at least once to have
obtained the page size.

put the call back where it originally was, and all should be well
again.