add support for aligned_alloc alignments 1M and over
the 16-bit offset to group header can't represent offsets over 1M in
16-byte units. this prevented aligned_alloc from satisfying alignment
requests over 1M, and combined with initial offset cycling within the
slot by the underlying malloc, made it difficult to handle the
boundary case of exactly 1M. therefore, aligned_alloc simply always
failed with ENOMEM for alignments over 512k.

despite there being strong reasons not to use huge alignments -- they
necessitate either wasting address space for the lifetime of the
allocation, or doing things that can lead to catastrophic cascading
fragmentation -- some programs do indeed expect alignments of 1M and
larger to work. so support needs to be implemented.

this patch extends the in-band slot header to allow a 32-bit offset
just below the normal 4-byte header if the [-4] byte is nonzero and
the normal 16-bit offset is 0. validity of the group header obtained,
as well as the offset, is still checked as usual. there should be no
extra cost in get_meta since the branch on p[-4] replaces an existing
assertion. and the loss of the assertion should not significantly
impact overflow trapping, since an offset that validates with p[-4]!=0
is more than 0xffff and necessarily mismatches the correct offset to
the group header.