consolidate non-portable system glue in a new header glue.h
this is intended to simplify upstreaming in musl (where the version of
glue.h here can be completely discarded and replaced by appropriate
libc glue) as well as facilitating build on systems that might lack
one or more Linux-specific interfaces. fallbacks are added for some of
these already where they can be detected by macros, but others
(especially getentropy) are still used unconditionally. this will be
improved later.