add support for allocating meta areas via legacy brk
this idea came out of a discussion with Daniel Micay looking for a way
to randomize the location of meta areas independently of the ASLR
offset where mmap starts, and particularly to avoid allocating new
ones in regions that previously contained memory belonging to the
application.

with kernel.randomize_va_space=2 (the default on modern kernels), the
legacy brk area generally fills these requirements. it's positioned at
a decent-sized random offset above the main program's text/data
mappings, and mmap aims to avoid this area so as not to break
traditional malloc implementations that rely on ability to expand the
brk area. moreover, since it expands contiguously for as long as
possible (then starts failing if it can't), we don't need to waste
VMAs on separate guard pages each time more meta slots are needed.

since use of brk is not portable and not necessarily safe outside of
libc, -DUSE_BRK has to be added to CFLAGS (e.g. from config.mak) to
enable it at build time. (for integration in musl this will be handled
by the replacement glue header.)

if brk fails, either due to not being supported at build time or run
time, or due to running into another mapping, the existing mmap-based
meta area allocation logic is used, and meta areas are protected by
new guard pages each time, just like before.