count slots that aren't available but free in coarse classing decision
under just sequential allocation with existing behavior, slots in the
coarse-sized class that haven't yet been filled will all be available
(active). however, if allocations and frees have already happened in
the coarse class, it may be that the only remaining slots are in the
"freed" state not the "available" state. these should still be
preferred over allocating a new group.

in the future this may also affect incremental activation of slots in
new groups, even when the allocation pattern is purely sequential.

note that the access to freed_mask is unsynchronized, but the value of
the combined avail_mask and freed_mask cannot change between zero and
nonzero without holding a lock, and at worst reading a stale value
changes whether the coarse class is used one more time.