move logic for skipping fully-free groups to common code path
this was added in commit 150de6e89459c0f77b5555e509a5c693591a156b to
keep fully-freed groups untouched until they're actually needed.
however it was only applied to the try_avail code path where the
previously active group was being cycled to the end of the list; if
the previously active group filled up, it was possible to select the
fully-free group as the new active group.

note: use of the word "the" is justified because presently there is at
most one fully-free group in any given class.